Moved EditActivity class to proper subdirectory and added navigation from StackActivity to EditActivity
