+++ Stackd/.idea/modules.xml
@@ -3,8 +3,6 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/Stackd.iml" filepath="$PROJECT_DIR$/Stackd.iml" />
-      <module fileurl="file://$PROJECT_DIR$/Stackd.iml" filepath="$PROJECT_DIR$/Stackd.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
