+++ Stackd/app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
             </intent-filter>
         </activity>
         <activity android:name=".activities.StackActivity"></activity>
-        <activity android:name="com.stackd.stackd.EditActivity"></activity>
+        <activity android:name="com.stackd.stackd.activities.EditActivity"></activity>
     </application>
 
 </manifest>
