+++ Stackd/app/src/main/java/com/stackd/stackd/activities/EditActivity.java
@@ -1,4 +1,4 @@
-package com.stackd.stackd;
+package com.stackd.stackd.activities;
 
 import android.graphics.Color;
 import android.support.v7.app.AppCompatActivity;
@@ -6,6 +6,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.CheckBox;
 import android.widget.LinearLayout;
+
+import com.stackd.stackd.R;
 import com.stackd.stackd.helpers.Resume;
 import java.text.DateFormat;
