+++ Stackd/app/src/main/java/com/stackd/stackd/activities/StackActivity.java
@@ -1,6 +1,7 @@
 package com.stackd.stackd.activities;
 
 import android.content.DialogInterface;
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
@@ -29,8 +30,8 @@ public class StackActivity extends AppCompatActivity {
             public void onItemClick(AdapterView<?> parent, View v,
                                     int position, long id) {
                 // dummy result, should open the review activity for the given resume
-                Toast.makeText(StackActivity.this, "" + position,
-                        Toast.LENGTH_SHORT).show();
+                Intent i = new Intent(StackActivity.this, EditActivity.class);
+                startActivity(i);
             }
         });
     }
