+++ Stackd/app/src/main/res/layout/activity_edit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.stackd.stackd.EditActivity">
+    tools:context="com.stackd.stackd.activities.EditActivity">
 
     <ImageView
         android:id="@+id/imageView"
