+++ Stackd/app/src/main/res/layout/activity_stack.xml
@@ -15,7 +15,8 @@
 
         <HorizontalScrollView
             android:layout_width="fill_parent"
-            android:layout_height="wrap_content">
+            android:layout_height="wrap_content"
+            android:scrollbars="none">
             <LinearLayout
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
